FIX CONFLICT MERGE longer to master

+++ src/SalesManager.java
@@ -17,9 +17,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = Integer.MAX_VALUE;
-        for (int sale : sales) {
+    public long min() {
+        long min = Long.MAX_VALUE;
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -27,7 +27,7 @@ public class SalesManager {
         return min;
     }
 
-    public int average() {
+    public long average() {
         return (Arrays.stream(sales).sum() - min() - max()) /
                 (sales.length - 2);
     }
